Resource moved from monad to typeclasses
